Move tooltip to seperate file
